# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JagoEvent Ticketing API",
        "description": "REST API untuk aplikasi JagoEvent — marketplace tiket event. Digunakan oleh aplikasi mobile Flutter untuk browse events, order tickets, manage vendor, dan validasi tiket.",
        "contact": {
            "name": "JagoEvent Support",
            "email": "support@JagoEvent.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "JagoEvent API"
        }
    ],
    "paths": {
...
        "/tickets/user/{id}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Get semua tiket milik vendor (by user ID)",
                "operationId": "945df146f85ef537b4fd41fd28cdd7ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID (vendor)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar tiket",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "sku_id": {
                                                        "type": "integer"
                                                    },
                                                    "event_id": {
                                                        "type": "integer"
                                                    },
                                                    "ticket_code": {
                                                        "type": "string",
                                                        "example": "TKT20260001"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "enum": [
                                                            "available",
                                                            "booked",
                                                            "sold",
                                                            "redeem"
                                                        ]
                                                    },
                                                    "sku": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "price": {
                                                                "type": "integer"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "event": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/check-ticket": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Validasi & redeem tiket",
                "description": "Cek apakah ticket_code valid dan statusnya available. Jika valid, status diubah ke redeem.",
                "operationId": "88d61ea58e101d902ae90804494a5ee0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ticket_code"
                                ],
                                "properties": {
                                    "ticket_code": {
                                        "type": "string",
                                        "example": "TKT20260001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tiket valid & berhasil di-redeem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ticket redeemed successfully"
                                        },
                                        "isValid": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Tiket sudah di-redeem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ticket already redeemed"
                                        },
                                        "isValid": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tiket tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ticket not found"
                                        },
                                        "isValid": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your Sanctum token (without Bearer prefix)"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}